refactor(layout): extract shared sizing helpers into sizing.ts
